chore: fix stuff
