+++ packages/js-dpp/lib/stateTransition/StateTransitionFacade.js
@@ -83,6 +83,7 @@ const validateProtocolVersionFactory = require('../version/validateProtocolVersi
 const validateDataContractUpdateTransitionBasicFactory = require('../dataContract/stateTransition/DataContractUpdateTransition/validation/basic/validateDataContractUpdateTransitionBasicFactory');
 const validateDataContractUpdateTransitionStateFactory = require('../dataContract/stateTransition/DataContractUpdateTransition/validation/state/validateDataContractUpdateTransitionStateFactory');
 const validateIndicesAreBackwardCompatible = require('../dataContract/stateTransition/DataContractUpdateTransition/validation/basic/validateIndicesAreBackwardCompatible');
+const getPropertyDefinitionByPath = require("../dataContract/getPropertyDefinitionByPath");
 
 class StateTransitionFacade {
   /**
@@ -107,8 +108,8 @@ class StateTransitionFacade {
       validateDataContractMaxDepth,
       enrichDataContractWithBaseSchema,
       validateDataContractPatterns,
-      RE2,
       validateProtocolVersion,
+      getPropertyDefinitionByPath,
     );
 
     const validateIdentityExistence = validateIdentityExistenceFactory(this.stateRepository);
